feat:library updated to check file timestamp of blob and missing blobs(RDCC-1927)

+++ src/functionalTest/java/uk/gov/hmcts/reform/locationrefdata/cameltest/testsupport/LrdBlobSupport.java
@@ -1,30 +1,34 @@
 package uk.gov.hmcts.reform.locationrefdata.cameltest.testsupport;
 
 import com.microsoft.azure.storage.CloudStorageAccount;
-import com.microsoft.azure.storage.StorageCredentialsAccountAndKey;
 import com.microsoft.azure.storage.blob.CloudBlobClient;
 import com.microsoft.azure.storage.blob.CloudBlobContainer;
 import com.microsoft.azure.storage.blob.CloudBlockBlob;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.ConfigFileApplicationContextInitializer;
 import org.springframework.stereotype.Component;
 import org.springframework.test.context.ContextConfiguration;
 import uk.gov.hmcts.reform.data.ingestion.configuration.AzureBlobConfig;
-import uk.gov.hmcts.reform.data.ingestion.configuration.StorageCredentials;
+import uk.gov.hmcts.reform.data.ingestion.configuration.BlobStorageCredentials;
 
 import java.io.InputStream;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import javax.annotation.PostConstruct;
-import javax.annotation.Resource;
 
 import static com.microsoft.azure.storage.blob.DeleteSnapshotsOption.INCLUDE_SNAPSHOTS;
+import static java.util.Objects.isNull;
 
 @Component
 @ContextConfiguration(classes = {
-    AzureBlobConfig.class, StorageCredentials.class}, initializers = ConfigFileApplicationContextInitializer.class)
+    AzureBlobConfig.class, BlobStorageCredentials.class}, initializers = ConfigFileApplicationContextInitializer.class)
 public class LrdBlobSupport {
 
+
+    @Autowired
+    @Qualifier("credscloudStorageAccount")
     CloudStorageAccount acc;
 
     CloudBlobClient cloudBlobClient;
@@ -33,15 +37,11 @@ public class LrdBlobSupport {
 
     CloudBlobContainer cloudBlobArchContainer;
 
-    @Resource(name = "credsreg")
-    StorageCredentialsAccountAndKey storageCredentialsAccountAndKey;
-
     @Value("${archival-date-format}")
     private String archivalDateFormat;
 
     @PostConstruct
     public void init() throws Exception {
-        acc = new CloudStorageAccount(storageCredentialsAccountAndKey, true);
         cloudBlobClient = acc.createCloudBlobClient();
         cloudBlobContainer = cloudBlobClient.getContainerReference("lrd-ref-data");
         cloudBlobArchContainer = cloudBlobClient.getContainerReference("lrd-ref-data-archive");
@@ -52,12 +52,15 @@ public class LrdBlobSupport {
         cloudBlockBlob.upload(sourceFile, 8 * 1024 * 1024);
     }
 
-    public void deleteBlob(String blob) throws Exception {
+    public void deleteBlob(String blob, boolean... status) throws Exception {
         CloudBlockBlob cloudBlockBlob = cloudBlobContainer.getBlockBlobReference(blob);
         cloudBlockBlob.delete(INCLUDE_SNAPSHOTS, null, null, null);
-
         String date = new SimpleDateFormat(archivalDateFormat).format(new Date());
-        cloudBlockBlob = cloudBlobArchContainer.getBlockBlobReference(blob.concat(date));
-        cloudBlockBlob.delete(INCLUDE_SNAPSHOTS, null, null, null);
+
+        //Skipped for Stale non existing files as not archived
+        if (isNull(status)) {
+            cloudBlockBlob = cloudBlobArchContainer.getBlockBlobReference(blob.concat(date));
+            cloudBlockBlob.delete(INCLUDE_SNAPSHOTS, null, null, null);
+        }
     }
 }
